Title: 계단 오르기, Time: 6ms, Memory: 0MB, Status: Wrong Answer - CodeTree

+++ 231120/계단 오르기/climbing-stairs.cpp
@@ -6,9 +6,9 @@
 using namespace std;
 
 // 변수 선언
-int n;
+long long n;
 
-int dp[MAX_N + 1];
+long long dp[MAX_N + 1];
 
 int main() {
     // 입력:
